requirements.txt
altgraph==0.17.4
asgiref==3.8.1
autopep8==1.5.7
bottle==0.13.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
dj-database-url==0.5.0
Django==3.2.6
django-environ==0.11.2
django-suit==0.2.28
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.8
joblib==1.4.2
mysqlclient==2.2.4
nltk==3.9.1
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pefile==2024.8.26
proxy_tools==0.1.0
pycodestyle==2.7.0
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-decouple==3.4
pythonnet==3.0.3
pytz==2021.1
pywebview==5.2
pywin32-ctypes==0.2.3
regex==2024.7.24
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
shiboken6==6.7.2
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
Unipath==1.1
urllib3==2.2.2
vaderSentiment==3.3.2
vboxapi==1.0
whitenoise==5.3.0
